+++ next.config.mjs
@@ -18,9 +18,6 @@ const nextConfig = {
       },
     ],
   },
-  experimental: {
-    appDir: true,
-  },
   compiler: {
     styledComponents: true,
   },
